requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beanie==1.26.0
boto3==1.35.22
botocore==1.35.22
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
jmespath==1.0.1
lazy-model==0.2.0
motor==3.6.0
openai==1.46.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dateutil==2.9.0.post0
regex==2024.9.11
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tiktoken==0.7.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3
uvicorn==0.30.6
python-dotenv
pytest
pytest-asyncio
mongomock-motor